[chore] : components 폴더 내부 파일 포맷 변경
