+++ components/AutoToast.jsx
@@ -80,3 +80,5 @@ function EmergencyAlert({ message, type }) {
     </div>
   )
 }
+
+
