+++ components/ButtonKakaoLogin.jsx
@@ -24,3 +24,5 @@ export default function ButtonKakaoLogin({ isLoading, onClick, disabled }) {
     </button>
   );
 }
+
+
