--- libs/api/supabase.ts
@@ -1,3 +1,0 @@
-import 'client-only'
-// 클라이언트 전용 파일로 변경 - 서버 함수들은 libs/api/server.ts로 이동
-export { createClient } from './client';
